+++ codex/settings/__init__.py
@@ -41,6 +41,7 @@ BUILD = not_falsy_env("BUILD")
 SLOW_QUERY_LIMIT = float(environ.get("CODEX_SLOW_QUERY_LIMIT", "0.5"))
 LOG_RESPONSE_TIME = not_falsy_env("CODEX_LOG_RESPONSE_TIME")
 LOG_REQUEST = not_falsy_env("CODEX_LOG_REQUEST")
+LOG_AUTH_HEADERS = not_falsy_env("CODEX_LOG_AUTH_HEADERS")
 # Misc
 VITE_HOST = environ.get("VITE_HOST")
 LOG_RETENTION = environ.get("LOG_RETENTION", "6 months")
@@ -70,6 +71,7 @@ RESET_ADMIN = not_falsy_env("CODEX_RESET_ADMIN")
 LOG_DIR = Path(environ.get("CODEX_LOG_DIR", CONFIG_PATH / "logs"))
 LOG_TO_CONSOLE = environ.get("CODEX_LOG_TO_CONSOLE") != "0"
 LOG_TO_FILE = environ.get("CODEX_LOG_TO_FILE") != "0"
+AUTH_REMOTE_USER = not_falsy_env("CODEX_AUTH_REMOTE_USER")
 THROTTLE_ANON = int(environ.get("CODEX_THROTTLE_ANON", "0"))
 THROTTLE_USER = int(environ.get("CODEX_THROTTLE_USER", "0"))
 THROTTLE_OPDS = int(environ.get("CODEX_THROTTLE_OPDS", "0"))
@@ -149,6 +151,7 @@ def _get_installed_apps():
         "whitenoise.runserver_nostatic",
         "django.contrib.staticfiles",
         "rest_framework",
+        "rest_framework.authtoken",
         "rest_registration",
         "corsheaders",
     ]
@@ -176,6 +179,10 @@ def _get_middleware():
         "django.middleware.common.CommonMiddleware",
         "django.middleware.csrf.CsrfViewMiddleware",
         "django.contrib.auth.middleware.AuthenticationMiddleware",
+    ]
+    if AUTH_REMOTE_USER:
+        middleware += ["django.contrib.auth.middleware.RemoteUserMiddleware"]
+    middleware += [
         "django.contrib.messages.middleware.MessageMiddleware",
         "django.middleware.clickjacking.XFrameOptionsMiddleware",
         "codex.middleware.TimezoneMiddleware",
@@ -254,6 +261,12 @@ SESSION_ENGINE = "django.contrib.sessions.backends.cached_db"
 #   Django 4.0 at the earliest:
 #   https://code.djangoproject.com/ticket/32674
 DEFAULT_AUTO_FIELD = "django.db.models.AutoField"
+
+if AUTH_REMOTE_USER:
+    AUTHENTICATION_BACKENDS = [
+        "django.contrib.auth.backends.RemoteUserBackend",
+    ]
+    logger.info("Remote User authorization enabled.")
 
 
 # Password validation
@@ -339,6 +352,7 @@ for scope, value in _THROTTLE_MAP.items():
 REST_FRAMEWORK = {
     "DEFAULT_AUTHENTICATION_CLASSES": (
         "rest_framework.authentication.SessionAuthentication",
+        "rest_framework.authentication.TokenAuthentication",
     ),
     "DEFAULT_RENDERER_CLASSES": (
         "djangorestframework_camel_case.render.CamelCaseJSONRenderer",