+++ codex/startup/__init__.py
@@ -7,6 +7,7 @@ from django.core.cache import cache
 from django.db.models import F, Q
 from django.db.models.functions import Now
 from loguru import logger
+from rest_framework.authtoken.models import Token
 
 from codex.choices.admin import AdminFlagChoices
 from codex.librarian.status_controller import STATUS_DEFAULTS
@@ -156,6 +157,16 @@ def update_custom_covers_for_config_dir():
         )
 
 
+def create_missing_auth_tokens():
+    """Create missing auth tokens."""
+    num_created = 0
+    for user in User.objects.all():
+        _, created = Token.objects.get_or_create(user=user)
+        if created:
+            num_created += 1
+        logger.info(f"Created {num_created} missing auth tokens for users.")
+
+
 def ensure_db_rows():
     """Ensure database content is good."""
     ensure_superuser()
@@ -165,6 +176,7 @@ def ensure_db_rows():
     clear_library_status()
     init_custom_cover_dir()
     update_custom_covers_for_config_dir()
+    create_missing_auth_tokens()
 
 
 def codex_init():