+++ codex/views/opds/v1/facets.py
@@ -7,7 +7,6 @@ from django.urls import reverse
 
 from codex.choices.admin import AdminFlagChoices
 from codex.models import AdminFlag
-from codex.views.browser.browser import BrowserView
 from codex.views.opds.const import MimeType, Rel, UserAgentNames
 from codex.views.opds.feed import OPDSBrowserView
 from codex.views.opds.util import get_user_agent_name
@@ -20,15 +19,13 @@ from codex.views.opds.v1.const import (
     RootFacetGroups,
 )
 from codex.views.opds.v1.entry.entry import OPDS1Entry
+from codex.views.template import CodexXMLTemplateMixin
 from codex.views.util import pop_name
 
 
-class OPDS1FacetsView(OPDSBrowserView):
+class OPDS1FacetsView(CodexXMLTemplateMixin, OPDSBrowserView):
     """OPDS 1 Facets methods."""
 
-    DEFAULT_ROUTE = MappingProxyType(
-        {**BrowserView.DEFAULT_ROUTE, "name": "opds:v1:feed"}
-    )
     TARGET = "opds1"
 
     def __init__(self, *args, **kwargs):