+++ NEWS.md
@@ -6,6 +6,19 @@ width: 128px;
 border-radius: 128px;
 " />
 
+## v1.9.8
+
+- Fixes
+    - Major Access Control Bug would let users see comics in include groups they
+      were not a part of.
+    - Remove redundant top links from OPDS 2.0 facets group block.
+    - Fixed a status crash.
+- Features
+    - Remote-User HTTP header SSO support activated by an environment variable.
+      See docs.
+    - HTTP Header Auth Tokens are now accepted as authorization. Useful for OPDS
+      and possibly some SSO setups.
+
 ## v1.9.7
 
 - Fixes